refactor(api): split compensate mapping into compensation module
action.rs was 295 lines (5 below the 300-line cap) and combined the
Action enum definition with names, capabilities, summaries, and
compensation metadata. Move the compensate / is_reversible mapping into
a sibling 'compensation' module that adds an inherent impl block on
Action. Behavior is unchanged (all 10 existing tests still pass) and
action.rs drops to 245 lines, leaving headroom for the next batch of
actions without bumping the cap.

tracks: Te72b7cc9-863b-4b41-9ff6-0af889c8c237
