+++ crates/convergio-api/src/action.rs
@@ -242,54 +242,4 @@ impl Action {
             Self::AgentPrompt => "Return the canonical agent prompt addendum.",
         }
     }
-
-    /// Compensating action that undoes this one's side effects when
-    /// a clean inverse exists (P3-3 — Palantir-inspired compensating
-    /// actions). Returns `None` for actions whose effect cannot be
-    /// reversed by another action of the same surface.
-    pub fn compensate(self) -> Option<Self> {
-        match self {
-            Self::RegisterAgent => Some(Self::RetireAgent),
-            Self::RetireAgent => Some(Self::RegisterAgent),
-            Self::ClaimWorkspaceLease => Some(Self::ReleaseWorkspaceLease),
-            Self::Status
-            | Self::CreatePlan
-            | Self::CreateTask
-            | Self::ListTasks
-            | Self::NextTask
-            | Self::ClaimTask
-            | Self::Heartbeat
-            | Self::AddEvidence
-            | Self::GetTaskContext
-            | Self::PublishMessage
-            | Self::PollMessages
-            | Self::AckMessage
-            | Self::SubmitTask
-            | Self::ValidatePlan
-            | Self::AuditVerify
-            | Self::ImportCrdtOps
-            | Self::ListCrdtConflicts
-            | Self::ListAgents
-            | Self::HeartbeatAgent
-            | Self::SpawnRunner
-            | Self::PlannerSolve
-            | Self::ListCapabilities
-            | Self::GetCapability
-            | Self::ListWorkspaceLeases
-            | Self::ReleaseWorkspaceLease
-            | Self::SubmitPatchProposal
-            | Self::EnqueuePatchProposal
-            | Self::ProcessMergeQueue
-            | Self::ListMergeQueue
-            | Self::ListWorkspaceConflicts
-            | Self::ExplainLastRefusal
-            | Self::AgentPrompt
-            | Self::AuditAppend => None,
-        }
-    }
-
-    /// True iff [`Self::compensate`] returns `Some`.
-    pub fn is_reversible(self) -> bool {
-        self.compensate().is_some()
-    }
 }
